refactor: rework editor and readonly editor layers
- de-couple readonly editor from platecontainer / editor toolbar wrappers

Re-works the PlateContainer and EditorToolbar wrappers to live alongside and within the /editor set of components to which their scoped, both because the readonly editor does not use them, and to pave way for a markdown (text) editor that needs its own plate setup and potentially toolbar elements
